# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AIC-DCT API
  version: 0.1.0
  description: |
    AIC-DCT backend API contract. Some Phase 4A Subject endpoints are still
    planned and MUST be added here in the same change that implements them;
    this document must not claim that a planned route is already available.

    Authorization is permission plus scope. List, aggregate, and export
    endpoints resolve an authorized project scope; detail and mutation
    endpoints verify the target object's project and evaluate the actor's
    role in that target project. Subject endpoints additionally derive the
    bound subject/project from authenticated identity and never trust a
    client-supplied subjectId or projectId to widen access.

    AI adopt/reject endpoints do not by themselves make an AI workflow
    compliant. Any endpoint that promotes AI-derived content into a formal
    record, including AI-derived protocol activation and report confirmation,
    must accept only Adopted or EditedAdopted output and must verify project,
    output kind, target object, and version. Purely human-authored records use
    an explicit non-AI path.

    Subject-originated source data cannot be silently overwritten by Web
    users. Approved assisted entry is a distinct, attributed, reasoned,
    persistently labelled, audited record and is never represented as Subject
    self-entry.

    Every error response conforms to:
    ```
...
    ```
servers:
  - url: http://localhost:4000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/subject/me:
    get:
      summary: Subject identity profile (bound subject + project; M4A)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Subject context }
        '403': { description: Not a bound Subject identity }
  /api/subject/my-tasks:
    get:
      summary: Today's tasks for bound Subject (questionnaires, visits, consent)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Grouped task buckets }
  /api/subject/epro/responses:
    get:
      summary: List own ePRO responses (identity-scoped)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Subject-owned responses }
    post:
      summary: Start own ePRO response (SubjectSelfReport channel)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Created response }
  /api/subject/epro/responses/{responseId}/submit:
    parameters:
      - { name: responseId, in: path, required: true, schema: { type: string } }
    post:
      summary: Submit own ePRO response (immutable after submit except attributed correction)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Submitted }
  /api/subject/symptom-reports:
    post:
      summary: Submit symptom report; severe flags create risk + safety draft
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Report accepted with urgent-care guidance when severe }
...